ppi ordering
