+++ api/db-cdr/changelog-data/db.changelog-13.xml
@@ -33,17 +33,6 @@
       <column name="source_count_value" type="NUMERIC" />
     </loadData>
 
-    <loadData tableName="achilles_results" file="csv/achilles_results_3.csv" encoding="UTF-8" quotchar='"'>
-      <column name="analysis_id" type="NUMERIC"/>
-      <column name="stratum_1" type="STRING"/>
-      <column name="stratum_2" type="STRING"/>
-      <column name="stratum_3" type="STRING"/>
-      <column name="stratum_4" type="STRING"/>
-      <column name="stratum_5" type="STRING"/>
-      <column name="count_value" type="NUMERIC"/>
-      <column name="source_count_value" type="NUMERIC" />
-    </loadData>
-
   </changeSet>
 
 </databaseChangeLog>
