+++ api/db-cdr/csv/achilles_analysis.csv
@@ -8,7 +8,7 @@ analysis_id,"analysis_name","stratum_1_name","stratum_2_name","stratum_3_name","
 6,"Number of persons by age decile","age_decile","","","","","column","counts"
 10,"Number of all persons by year of birth by gender","year_of_birth","gender_concept_id","","","","column","counts"
 12,"Number of persons by race and ethnicity","race_concept_id","ethnicity_concept_id","","","","pie","counts"
-1814,"Distribution of numeric values by measurement concept id ","measurement_concept_id","unit_name",NULL,"","","box","distribution"
+1814,"Distribution of numeric values by measurement concept id ","measurement_concept_id","unit_name",NULL,NULL,NULL,"box","distribution"
 1815,"Distribution of numeric values, by measurement_concept_id and gender","measurement_concept_id","unit_name","gender_concept_id","IQR_min","IQR_max","box","distribution"
 1816,"Distribution of numeric values by measurement_concept_id and age at occurrence decile","measurement_concept_id","unit_name","age_decile","IQR_min","IQR_max","box","distribution"
 1900,"Measurement response gender histogram","measurement_concept_id","unit_name","gender_concept_id","value","value_as_concept_id","column","counts"
@@ -17,7 +17,7 @@ analysis_id,"analysis_name","stratum_1_name","stratum_2_name","stratum_3_name","
 3005,"Counts per person distribution by concept id","measurement_concept_id",NULL,NULL,NULL,NULL,"box","distribution"
 3006,"Counts per person by gender distribution","measurement_concept_id","gender_concept_id",NULL,NULL,NULL,"box","distribution"
 3007,"Counts per person by age at occurrence","measurement_concept_id","age_decile_at_occurrence",NULL,NULL,NULL,"box","distribution"
-3008,"Counts per person distrubtion by age at occurrence and gender","measurement_concept_id","age_decile_at_occurrence","gender_concept_id",NULL,NULL,"box","distribution"
+3008,"Counts per person distrubtion by age at occurrence and gender","measurement_concept_id","age_decile_at_occurrence","gender",NULL,NULL,"box","distribution"
 3101,"Gender","concept_id","ppi_sex_at_birth_concept_id","domain_id",NULL,NULL,"pie","counts"
 3102,"Age at Occurrence","concept_id","age_decile","domain_id","null","","column","counts"
 3103,"Race","concept_id","race_concept_id","domain_id",NULL,NULL,"column","counts"
