+++ api/db-cdr/csv/survey_module.csv
@@ -1,4 +1,4 @@
 "concept_id","name","description","question_count","participant_count"
-1585855,"Lifestyle","Lifestyle survey includes information on participant smoking, alcohol and recreational drug use.",20,568120
-1585710,"Overall Health","Overall Health survey provides information about how participants report levels of individual health.",12,567857
-1586134,"The Basics","The Basics survey includes participant demographic information.",23,567437
+1585855,"Lifestyle","Lifestyle survey includes information on participant smoking, alcohol and recreational drug use.",28,568120
+1585710,"Overall Health","Overall Health survey provides information about how participants report levels of individual health.",24,567857
+1586134,"The Basics","The Basics survey includes participant demographic information.",26,567437
