+++ api/db-cdr/generate-cdr/bq-schemas/survey_question_map.json
@@ -1,0 +1,17 @@
+[
+  {
+    "mode": "REQUIRED",
+    "name": "id",
+    "type": "INTEGER"
+  },
+  {
+    "mode": "REQUIRED",
+    "name": "survey_concept_id",
+    "type": "INTEGER"
+  },
+  {
+    "mode": "REQUIRED",
+    "name": "question_concept_id",
+    "type": "INTEGER"
+  }
+]
