+++ api/db-cdr/generate-cdr/make-bq-data.sh
@@ -94,7 +94,7 @@ fi
 
 # Create bq tables we have json schema for
 schema_path=generate-cdr/bq-schemas
-create_tables=(achilles_analysis achilles_results achilles_results_concept achilles_results_dist concept concept_relationship criteria criteria_attribute domain_info survey_module domain vocabulary concept_ancestor concept_synonym domain_vocabulary_info unit_map)
+create_tables=(achilles_analysis achilles_results achilles_results_concept achilles_results_dist concept concept_relationship criteria criteria_attribute domain_info survey_module domain vocabulary concept_ancestor concept_synonym domain_vocabulary_info unit_map survey_question_map)
 
 for t in "${create_tables[@]}"
 do
@@ -103,7 +103,7 @@ do
 done
 
 # Load tables from csvs we have. This is not cdr data but meta data needed for workbench app
-load_tables=(domain_info survey_module achilles_analysis unit_map)
+load_tables=(domain_info survey_module achilles_analysis unit_map survey_question_map)
 csv_path=generate-cdr/csv
 for t in "${load_tables[@]}"
 do
